Add paragraph character count in airline
Make a function to count characters in current paragraph, and surface it
as a plugin in airline that shows alongside the wordcount.

It looks like this:

    423 words  ¶ 132

Why would you want to surface paragraph character count? Because I
semi-frequently draft things (reviews, feedback) that have hard
character limits, and keeping an eye on them while I draft is handy.